Split FunctionResource to smaller controllers. FunctionResource, FunctionMvcResource, FunctionUploadResource, FunctionApiDocsResource
