+++ src/main/java/com/icthh/xm/ms/entity/web/rest/FunctionResource.java
@@ -5,11 +5,9 @@ import com.icthh.xm.commons.logging.aop.IgnoreLogginAspect;
 import com.icthh.xm.commons.permission.annotation.PrivilegeDescription;
 import com.icthh.xm.ms.entity.domain.FunctionContext;
 import com.icthh.xm.ms.entity.service.FunctionService;
-import com.icthh.xm.ms.entity.service.swagger.DynamicSwaggerFunctionGenerator;
+import com.icthh.xm.ms.entity.web.rest.util.FunctionUtils;
 import com.icthh.xm.ms.entity.web.rest.util.HeaderUtil;
-import io.undertow.util.BadRequestException;
 import jakarta.servlet.http.HttpServletRequest;
-import lombok.SneakyThrows;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
@@ -18,7 +16,6 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
-import org.springframework.util.AntPathMatcher;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PatchMapping;
@@ -29,19 +26,12 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
-import org.springframework.web.multipart.MultipartFile;
-import org.springframework.web.servlet.ModelAndView;
 
 import java.net.URI;
-import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 
-import static com.google.common.collect.ImmutableMap.of;
-import static com.icthh.xm.ms.entity.config.Constants.MVC_FUNC_RESULT;
 import static com.icthh.xm.ms.entity.service.impl.FunctionServiceImpl.POST_URLENCODED;
-import static org.springframework.web.servlet.HandlerMapping.BEST_MATCHING_PATTERN_ATTRIBUTE;
-import static org.springframework.web.servlet.HandlerMapping.PATH_WITHIN_HANDLER_MAPPING_ATTRIBUTE;
 
 /**
  * The {@link FunctionResource} class.
@@ -55,30 +45,17 @@ public class FunctionResource {
 
     private static final String ENTITY_NAME_FUNCTION_PATH = "/api/function-contexts/";
 
-    private static final String UPLOAD = "/upload";
-
     private final FunctionService functionService;
-    private final DynamicSwaggerFunctionGenerator functionDocService;
 
     private FunctionResource self;
 
-    public FunctionResource(@Qualifier("functionService") FunctionService functionService,
-                            DynamicSwaggerFunctionGenerator functionDocService) {
+    public FunctionResource(@Qualifier("functionService") FunctionService functionService) {
         this.functionService = functionService;
-        this.functionDocService = functionDocService;
     }
 
     @Autowired
     public void setSelf(@Lazy FunctionResource self) {
         this.self = self;
-    }
-
-    @Timed
-    @GetMapping("/functions/api-docs")
-    @PrivilegeDescription("Privilege to get openapi documentation for functions api")
-    public ResponseEntity<Object> callGetFunction(HttpServletRequest request) {
-        String url = "https://" + request.getHeader("x-domain") + ":" + request.getHeader("x-port");
-        return ResponseEntity.ok().body(functionDocService.generateSwagger(url));
     }
 
     @Timed
@@ -203,66 +180,13 @@ public class FunctionResource {
         return ResponseEntity.ok().body(result.functionResult());
     }
 
-    /**
-     * POST  /functions/mvc/{functionKey} : Execute a mvc function by key (key in entity specification).
-     *
-     * @param functionKey   the function key to execute
-     * @param functionInput function input data context
-     * @return ModelAndView object
-     */
-    @Timed
-    @PostMapping("/functions/mvc/{functionKey:.+}")
-    @PreAuthorize("hasPermission({'functionKey': #functionKey}, 'FUNCTION.MVC.CALL')")
-    @PrivilegeDescription("Privilege to execute a mvc function by key (key in entity specification)")
-    public ModelAndView callMvcFunction(@PathVariable("functionKey") String functionKey,
-                                        @RequestBody(required = false) Map<String, Object> functionInput) {
-        FunctionContext result = functionService.execute(functionKey, functionInput, "POST");
-        return getMvcResult(result);
-    }
-
-    @Timed
-    @PostMapping(value = "/functions/mvc/{functionKey:.+}", consumes = {
-        MediaType.APPLICATION_FORM_URLENCODED_VALUE,
-    })
-    @PreAuthorize("hasPermission({'functionKey': #functionKey}, 'FUNCTION.MVC.CALL')")
-    @PrivilegeDescription("Privilege to execute a mvc function by key (key in entity specification)")
-    public ModelAndView callPostFormMvcFunction(@PathVariable("functionKey") String functionKey,
-                                                @RequestParam(required = false) Map<String, Object> functionInput) {
-        return callMvcFunction(functionKey, functionInput);
-    }
-
-    @Timed
-    @PostMapping("/functions/anonymous/mvc/{functionKey:.+}")
-    public ModelAndView callMvcAnonymousFunction(@PathVariable("functionKey") String functionKey,
-                                                 @RequestBody(required = false) Map<String, Object> functionInput) {
-        FunctionContext result = functionService.executeAnonymous(functionKey, functionInput, "POST");
-        return getMvcResult(result);
-    }
-
-    @Timed
-    @GetMapping("/functions/anonymous/mvc/{functionKey:.+}")
-    public ModelAndView callMvcGetAnonymousFunction(@PathVariable("functionKey") String functionKey,
-                                                 @RequestParam(required = false) Map<String, Object> functionInput) {
-        FunctionContext result = functionService.executeAnonymous(functionKey, functionInput, "GET");
-        return getMvcResult(result);
-    }
-
-    @Timed
-    @PostMapping(value = "/functions/anonymous/mvc/{functionKey:.+}", consumes = {
-        MediaType.APPLICATION_FORM_URLENCODED_VALUE,
-    })
-    public ModelAndView callPostFormMvcAnonymousFunction(@PathVariable("functionKey") String functionKey,
-                                                         @RequestParam(required = false) Map<String, Object> functionInput) {
-        FunctionContext result = functionService.executeAnonymous(functionKey, functionInput, POST_URLENCODED);
-        return getMvcResult(result);
-    }
-
     @IgnoreLogginAspect
     @Timed
     @GetMapping("/functions/**")
     public ResponseEntity<Object> callGetFunction(HttpServletRequest request,
                                   @RequestParam(required = false) Map<String, Object> functionInput) {
-        return self.callGetFunction(getFunctionKey(request), functionInput);
+        String functionKey = FunctionUtils.getFunctionKey(request);
+        return self.callGetFunction(functionKey, functionInput);
     }
 
     @IgnoreLogginAspect
@@ -270,7 +194,8 @@ public class FunctionResource {
     @PostMapping(value = "/functions/**")
     public ResponseEntity<Object> callPostFunction(HttpServletRequest request,
                                                    @RequestBody(required = false) Map<String, Object> functionInput) {
-        return self.callPostFunction(getFunctionKey(request), functionInput);
+        String functionKey = FunctionUtils.getFunctionKey(request);
+        return self.callPostFunction(functionKey, functionInput);
     }
 
     @IgnoreLogginAspect
@@ -280,7 +205,8 @@ public class FunctionResource {
     })
     public ResponseEntity<Object> callPostFormFunction(HttpServletRequest request,
                                                        @RequestParam(required = false) Map<String, Object> functionInput) {
-        return self.callPostFormFunction(getFunctionKey(request), functionInput);
+        String functionKey = FunctionUtils.getFunctionKey(request);
+        return self.callPostFormFunction(functionKey, functionInput);
     }
 
     @IgnoreLogginAspect
@@ -288,7 +214,8 @@ public class FunctionResource {
     @PutMapping("/functions/**")
     public ResponseEntity<Object> callPutFunction(HttpServletRequest request,
                                                   @RequestBody(required = false) Map<String, Object> functionInput) {
-        return self.callPutFunction(getFunctionKey(request), functionInput);
+        String functionKey = FunctionUtils.getFunctionKey(request);
+        return self.callPutFunction(functionKey, functionInput);
     }
 
     @IgnoreLogginAspect
@@ -296,7 +223,8 @@ public class FunctionResource {
     @PatchMapping("/functions/**")
     public ResponseEntity<Object> callPatchFunction(HttpServletRequest request,
                                                     @RequestBody(required = false) Map<String, Object> functionInput) {
-        return self.callPatchFunction(getFunctionKey(request), functionInput);
+        String functionKey = FunctionUtils.getFunctionKey(request);
+        return self.callPatchFunction(functionKey, functionInput);
     }
 
     @IgnoreLogginAspect
@@ -304,51 +232,8 @@ public class FunctionResource {
     @DeleteMapping("/functions/**")
     public ResponseEntity<Object> callDeleteFunction(HttpServletRequest request,
                                                   @RequestBody(required = false) Map<String, Object> functionInput) {
-        return self.callDeleteFunction(getFunctionKey(request), functionInput);
-    }
-
-    @IgnoreLogginAspect
-    @Timed
-    @PostMapping("/functions/mvc/**")
-    public ModelAndView callMvcFunction(HttpServletRequest request,
-                                        @RequestBody(required = false) Map<String, Object> functionInput) {
-        return self.callMvcFunction(getFunctionKey(request), functionInput);
-    }
-
-    @IgnoreLogginAspect
-    @Timed
-    @PostMapping(value = "/functions/mvc/**", consumes = {
-        MediaType.APPLICATION_FORM_URLENCODED_VALUE,
-    })
-    public ModelAndView callPostFormMvcFunction(HttpServletRequest request,
-                                                @RequestParam(required = false) Map<String, Object> functionInput) {
-        return self.callMvcFunction(getFunctionKey(request), functionInput);
-    }
-
-    @IgnoreLogginAspect
-    @Timed
-    @GetMapping("/functions/anonymous/mvc/**")
-    public ModelAndView callMvcAnonymousGetFunction(HttpServletRequest request,
-                                                    @RequestParam(required = false) Map<String, Object> functionInput) {
-        return self.callMvcGetAnonymousFunction(getFunctionKey(request), functionInput);
-    }
-
-    @IgnoreLogginAspect
-    @Timed
-    @PostMapping("/functions/anonymous/mvc/**")
-    public ModelAndView callMvcAnonymousFunction(HttpServletRequest request,
-                                                 @RequestBody(required = false) Map<String, Object> functionInput) {
-        return self.callMvcAnonymousFunction(getFunctionKey(request), functionInput);
-    }
-
-    @IgnoreLogginAspect
-    @Timed
-    @PostMapping(value = "/functions/anonymous/mvc/**", consumes = {
-        MediaType.APPLICATION_FORM_URLENCODED_VALUE,
-    })
-    public ModelAndView callPostFromMvcAnonymousFunction(HttpServletRequest request,
-                                                         @RequestParam(required = false) Map<String, Object> functionInput) {
-        return self.callMvcAnonymousFunction(getFunctionKey(request), functionInput);
+        String functionKey = FunctionUtils.getFunctionKey(request);
+        return self.callDeleteFunction(functionKey, functionInput);
     }
 
     @IgnoreLogginAspect
@@ -356,7 +241,8 @@ public class FunctionResource {
     @GetMapping("/functions/anonymous/**")
     public ResponseEntity<Object> callGetAnonymousFunction(HttpServletRequest request,
                                                   @RequestParam(required = false) Map<String, Object> functionInput) {
-        return self.callGetAnonymousFunction(getFunctionKey(request), functionInput);
+        String functionKey = FunctionUtils.getFunctionKey(request);
+        return self.callGetAnonymousFunction(functionKey, functionInput);
     }
 
     @IgnoreLogginAspect
@@ -364,7 +250,8 @@ public class FunctionResource {
     @PostMapping("/functions/anonymous/**")
     public ResponseEntity<Object> callPostAnonymousFunction(HttpServletRequest request,
                                                             @RequestBody(required = false) Map<String, Object> functionInput) {
-        return self.callPostAnonymousFunction(getFunctionKey(request), functionInput);
+        String functionKey = FunctionUtils.getFunctionKey(request);
+        return self.callPostAnonymousFunction(FunctionUtils.getFunctionKey(request), functionInput);
     }
 
     @IgnoreLogginAspect
@@ -374,31 +261,7 @@ public class FunctionResource {
     })
     public ResponseEntity<Object> callPostFormAnonymousFunction(HttpServletRequest request,
                                                                 @RequestParam(required = false) Map<String, Object> functionInput) {
-        return self.callPostFormAnonymousFunction(getFunctionKey(request), functionInput);
-    }
-
-    @Timed
-    @PostMapping(value = "/functions/**", consumes = MediaType.MULTIPART_FORM_DATA_VALUE)
-    @PreAuthorize("hasPermission({'functionKey': #functionKey}, 'FUNCTION.UPLOAD.CALL')")
-    @SneakyThrows
-    @PrivilegeDescription("Privilege to call upload function")
-    public ResponseEntity<Object> callUploadFunction(HttpServletRequest request,
-                                                     @RequestParam(value = "file", required = false) List<MultipartFile> files,
-                                                     HttpServletRequest httpServletRequest) {
-        if (!request.getRequestURI().endsWith(UPLOAD)) {
-            return ResponseEntity.badRequest().body("Invalid upload url");
-        }
-        Map<String, Object> functionInput = of("httpServletRequest", httpServletRequest, "files", files);
-        String functionKey = getFunctionKey(request);
-        functionKey = functionKey.substring(0, functionKey.length() - UPLOAD.length());
-        FunctionContext result = functionService.execute(functionKey, functionInput, "POST");
-        return ResponseEntity.ok().body(result.functionResult());
-    }
-
-    public static String getFunctionKey(HttpServletRequest request) {
-        String path = request.getAttribute(PATH_WITHIN_HANDLER_MAPPING_ATTRIBUTE).toString();
-        String bestMatchingPattern = request.getAttribute(BEST_MATCHING_PATTERN_ATTRIBUTE).toString();
-        return new AntPathMatcher().extractPathWithinPattern(bestMatchingPattern, path);
+        return self.callPostFormAnonymousFunction(FunctionUtils.getFunctionKey(request), functionInput);
     }
 
     private ResponseEntity<Object> processCreatedResponse(FunctionContext result) {
@@ -410,21 +273,6 @@ public class FunctionResource {
             .orElse(ResponseEntity
                 .status(HttpStatus.CREATED.value())
                 .body(result.functionResult()));
-    }
-
-    private ModelAndView getMvcResult(FunctionContext context) {
-        if (context == null || context.getData() == null) {
-            return null;
-        }
-
-        Object modelAndView = context.getData().get(MVC_FUNC_RESULT);
-        if (modelAndView instanceof ModelAndView) {
-            return (ModelAndView) modelAndView;
-        }
-
-        log.warn("Context did not contain {} or type (ModelAndView) mismatch, present keys {}", MVC_FUNC_RESULT, context.getData().keySet());
-
-        return null;
     }
 
 }
